+++ klum-ast/src/test/groovy/com/blackbuild/klum/ast/util/FactoryTest.groovy
@@ -319,4 +319,78 @@ import com.blackbuild.groovy.configdsl.transform.Key
         instance.inner.getClass() == ConcreteInner
     }
 
+    @Issue("359")
+    def "convert map before creating instance"() {
+        given:
+        createClass '''
+import com.blackbuild.klum.ast.util.KlumFactory
+
+@DSL class Person {
+    String firstName
+    String lastName
+ 
+    static class Factory extends KlumFactory.Unkeyed<Person> {
+        protected Factory() { super(Person) }
+        
+        @Override
+        Person FromMap(Map map) {
+            def transformedMap = map.collectEntries { k, v ->
+                // transform key from kebap to camel case
+                [(k as String).tokenize('-').collect { it.capitalize() }.join('').uncapitalize(), v]
+            }
+            return super.FromMap(transformedMap)
+        }
+    }
+} 
+'''
+        when:
+        def person = Person.Create.FromMap(['first-name': 'Klaus', 'last-name': 'Müller'])
+
+        then:
+        person.firstName == 'Klaus'
+        person.lastName == 'Müller'
+    }
+
+    @Issue("359")
+    def "convert map before creating nested instance"() {
+        given:
+        createClass '''
+import com.blackbuild.klum.ast.util.KlumFactory
+
+@DSL class Person {
+    String firstName
+    String lastName
+ 
+    Address address
+} 
+
+@DSL class Address {
+    String street
+    String city
+
+    static class Factory extends KlumFactory.Unkeyed<Address> {
+        protected Factory() { super(Address) }
+        
+        @Override
+        Address FromMap(Map map) {
+            def transformedMap = [
+                street: map.number ? "$map.street $map.number" as String : map.street, 
+                city: map.zip ? "$map.city $map.zip" as String : map.city
+            ]
+            return super.FromMap(transformedMap)
+        }
+    }
+}
+
+'''
+        when:
+        def person = Person.Create.FromMap(['firstName': 'Klaus', 'lastName': 'Müller', address: ['street': 'Hauptstraße', 'number': '12', 'city': 'München', 'zip': '80331']])
+
+        then:
+        person.firstName == 'Klaus'
+        person.lastName == 'Müller'
+        person.address.street == 'Hauptstraße 12'
+        person.address.city == 'München 80331'
+    }
+
 }